[scopes] refactor mapScopes
