+++ src/utils/pause/mapScopes/types.js
@@ -1,0 +1,7 @@
+// @flow
+
+export type GeneratedBindingLocation = {
+  name: string,
+  loc: BindingLocation,
+  desc: BindingContents | null
+};
